// projects/qa/src/app/models/Environment.ts
export class Environment {

  environmentId: number;
  name: string;
  description: string;
  environmentVersion: string;
  Product: string;
  Installed_release: number;
  Install_id: any;
  Env_name: string;
  State: string;
  Internal_URL: string;
  Env_type: string;
  Env_id: number;
  External_URL: string;
}